jwt, refactor
