+++ modules/services/searching/src/main/java/org/fao/geonet/searching/SecurityConfigurer.java
@@ -5,24 +5,20 @@
 
 package org.fao.geonet.searching;
 
-import org.fao.geonet.common.GnUserAuthentificationConverter;
-import org.springframework.context.annotation.Bean;
+import org.fao.geonet.common.TokenStoreConfig;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
-import org.springframework.security.oauth2.provider.token.DefaultAccessTokenConverter;
-import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
-import org.springframework.security.oauth2.provider.token.TokenStore;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
-import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
 import org.springframework.security.web.authentication.HttpStatusEntryPoint;
 
 @EnableResourceServer
 @Configuration
 @Order(4)
+@Import(TokenStoreConfig.class)
 public class SecurityConfigurer extends ResourceServerConfigurerAdapter {
 
   @Override
@@ -40,35 +36,6 @@ public class SecurityConfigurer extends ResourceServerConfigurerAdapter {
             .authenticationEntryPoint(
                 new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED))
         );
-  }
-
-
-  public TokenStore tokenStore() {
-    return new JwtTokenStore(accessTokenConverter());
-  }
-
-  /**
-   * Access token converter which set signing key.
-   */
-  @Bean
-  public JwtAccessTokenConverter accessTokenConverter() {
-    JwtAccessTokenConverter converter = new JwtAccessTokenConverter();
-    converter.setSigningKey("123");
-    DefaultAccessTokenConverter accessTokenConverter = new DefaultAccessTokenConverter();
-    accessTokenConverter.setUserTokenConverter(new GnUserAuthentificationConverter());
-    converter.setAccessTokenConverter(accessTokenConverter);
-    return converter;
-  }
-
-  /**
-   * A default token service using JWT token store.
-   */
-  @Bean
-  public DefaultTokenServices tokenServices() {
-    DefaultTokenServices defaultTokenServices = new DefaultTokenServices();
-    defaultTokenServices.setTokenStore(tokenStore());
-    defaultTokenServices.setSupportRefreshToken(true);
-    return defaultTokenServices;
   }
 }
